requirements.txt
astroid==2.0.4
confusable-homoglyphs==3.2.0
Django==1.11
django-bootstrap3==11.0.0
django-bootstrap4==0.0.7
django-registration==2.4.1
django-tinymce==2.7.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pylint==2.1.1
pytz==2018.5
six==1.11.0
typed-ast==1.1.0
wrapt==1.10.11
